+++ packages/table-render/src/table-render/table-column-render.tsx
@@ -149,6 +149,6 @@ export function renderTableColumn<T>(options: TableColumnOptions<T>, pageMode: '
     ...{
       key: options.key,
     },
-    render,
+    render: render as any,
   }
 }